+++ tests/MiniZinc.SolveTests/CommandTests.fs
@@ -8,44 +8,45 @@ module CommandTests =
     open MiniZinc.Command
     open Xunit
     
+    let mz = MiniZinc()
+    
     [<Fact>]
     let ``parse flag`` () =
         let cmd = "--a"
         let arg = Arg.parse(cmd)
-        arg.ToString().AssertEquals("--a")
+        arg.StringEquals("--a")
 
     [<Fact>]
     let ``parse assign equals`` () =
         let arg = Arg.parse "--model=xd.json"
-        arg.Flag.AssertEquals "--model"
-        arg.Value.AssertEquals "xd.json"
+        arg.Flag.StringEquals "--model"
+        arg.Value.StringEquals "xd.json"
         
     [<Fact>]
     let ``parse value`` () =
         let input = "asdfasdf"
         let arg = Arg.parse input 
-        arg.Value.AssertEquals input
-        arg.Flag.AssertEquals ""
+        arg.Value.StringEquals input
+        arg.Flag.StringEquals ""
         
     [<Fact>]
     let ``parse many`` () =
         let a = Args.parse "--count=1"
         let b = Arg.parse "--count"
-        let args = Args.Create(a, b, 1)
-        args.ToString().AssertEquals "--count=1 --count 1"
+        let args = Args.create(a, b, 1)
+        args.StringEquals "--count=1 --count 1"
 
     [<Theory>]
     [<InlineData("org.gecode.gecode")>]
     [<InlineData("org.chuffed.chuffed")>]
     let ``test solver installed`` id =
-        let solver =
-            MiniZinc.GetSolver id
+        let solver = mz.GetSolver id
         let sid = solver.Result.Value.Id
-        sid.AssertEquals id
-        
+        sid.StringEquals id
+         
     [<Fact>]
     let ``test minizinc version`` () =
         task {
-            let! version = MiniZinc.Version ()
-            assert (version <> "")
+            let! version = mz.Version()
+            version <> ""
         }